Develop - Created iotApp module
